rag-app/requirements.txt
endee
sentence-transformers
flask
flask-cors
python-dotenv
requests
PyPDF2
groq
beautifulsoup4
